+++ src/main/java/fortville/components/ExecuteWriteback.java
@@ -23,9 +23,9 @@ public class ExecuteWriteback {
 
     public void clock() {
         Opcode opcode = fetchDecodeBuffer.getOpcode();
-        short data1 = fetchDecodeBuffer.getData1();
-        short data2 = fetchDecodeBuffer.getData2();
-        short data3 = fetchDecodeBuffer.getData3();
+        int data1 = fetchDecodeBuffer.getData1();
+        int data2 = fetchDecodeBuffer.getData2();
+        int data3 = fetchDecodeBuffer.getData3();
         opcode.execute(data1, data2, data3, memory, display, registers);
     }
 }
