+++ src/main/java/fortville/components/FetchDecode.java
@@ -53,9 +53,9 @@ public class FetchDecode {
     }
 
     public void clock() {
-        short opcode = memory.loadInstruction();
-        byte firstNum = (byte)((opcode & 0xF000) >> 12);
-        byte lastNum = (byte)(opcode & 0x000F);
+        int opcode = memory.loadInstruction();
+        int firstNum = (opcode & 0xF000) >> 12;
+        int lastNum = opcode & 0x000F;
         if (opcode == 0) {
             // Tempory thing to end the program when it's done, will need to be changed
             System.exit(0);
@@ -75,83 +75,83 @@ public class FetchDecode {
             break;
         case 1:
             fetchDecodeBuffer.setOpcode(new JumpAddress());
-            fetchDecodeBuffer.setData1((short)(opcode & 0x0FFF));
+            fetchDecodeBuffer.setData1(opcode & 0x0FFF);
             break;
         case 2:
             fetchDecodeBuffer.setOpcode(new CallAddress());
-            fetchDecodeBuffer.setData1((short)(opcode & 0x0FFF));
+            fetchDecodeBuffer.setData1(opcode & 0x0FFF);
             break;
         case 3:
             fetchDecodeBuffer.setOpcode(new SkipNextInstructionEqualConst());
-            fetchDecodeBuffer.setData1((short)((opcode & 0x0F00) >> 8));
-            fetchDecodeBuffer.setData2((short)(opcode & 0x00FF));
+            fetchDecodeBuffer.setData1((opcode & 0x0F00) >> 8);
+            fetchDecodeBuffer.setData2(opcode & 0x00FF);
             break;
         case 4:
             fetchDecodeBuffer.setOpcode(new SkipNextInstructionNotEqualConst());
-            fetchDecodeBuffer.setData1((short)((opcode & 0x0F00) >> 8));
-            fetchDecodeBuffer.setData2((short)(opcode & 0x00FF));
+            fetchDecodeBuffer.setData1((opcode & 0x0F00) >> 8);
+            fetchDecodeBuffer.setData2(opcode & 0x00FF);
             break;
         case 5:
             fetchDecodeBuffer.setOpcode(new SkipNextInstructionEqualRegister());
-            fetchDecodeBuffer.setData1((short)((opcode & 0x0F00) >> 8));
-            fetchDecodeBuffer.setData2((short)((opcode & 0x00F0) >> 4));
+            fetchDecodeBuffer.setData1((opcode & 0x0F00) >> 8);
+            fetchDecodeBuffer.setData2((opcode & 0x00F0) >> 4);
             break;
         case 6:
             fetchDecodeBuffer.setOpcode(new SetRegisterEqualConst());
-            fetchDecodeBuffer.setData1((short)((opcode & 0x0F00) >> 8));
-            fetchDecodeBuffer.setData2((short)(opcode & 0x00FF));
+            fetchDecodeBuffer.setData1((opcode & 0x0F00) >> 8);
+            fetchDecodeBuffer.setData2(opcode & 0x00FF);
             break;
         case 7:
             fetchDecodeBuffer.setOpcode(new AddConstToRegister());
-            fetchDecodeBuffer.setData1((short)((opcode & 0x0F00) >> 8));
-            fetchDecodeBuffer.setData2((short)(opcode & 0x00FF));
+            fetchDecodeBuffer.setData1((opcode & 0x0F00) >> 8);
+            fetchDecodeBuffer.setData2(opcode & 0x00FF);
             break;
         case 8:
             switch (lastNum) {
             case 0:
                 fetchDecodeBuffer.setOpcode(new SetRegisterToRegister());
-                fetchDecodeBuffer.setData1((short)((opcode & 0x0F00) >> 8));
-                fetchDecodeBuffer.setData2((short)((opcode & 0x00F0) >> 4));
+                fetchDecodeBuffer.setData1((opcode & 0x0F00) >> 8);
+                fetchDecodeBuffer.setData2((opcode & 0x00F0) >> 4);
                 break;
             case 1:
                 fetchDecodeBuffer.setOpcode(new OrRegister());
-                fetchDecodeBuffer.setData1((short)((opcode & 0x0F00) >> 8));
-                fetchDecodeBuffer.setData2((short)((opcode & 0x00F0) >> 4));
+                fetchDecodeBuffer.setData1((opcode & 0x0F00) >> 8);
+                fetchDecodeBuffer.setData2((opcode & 0x00F0) >> 4);
                 break;
             case 2:
                 fetchDecodeBuffer.setOpcode(new AndRegister());
-                fetchDecodeBuffer.setData1((short)((opcode & 0x0F00) >> 8));
-                fetchDecodeBuffer.setData2((short)((opcode & 0x00F0) >> 4));
+                fetchDecodeBuffer.setData1((opcode & 0x0F00) >> 8);
+                fetchDecodeBuffer.setData2((opcode & 0x00F0) >> 4);
                 break;
             case 3:
                 fetchDecodeBuffer.setOpcode(new XORRegister());
-                fetchDecodeBuffer.setData1((short)((opcode & 0x0F00) >> 8));
-                fetchDecodeBuffer.setData2((short)((opcode & 0x00F0) >> 4));
+                fetchDecodeBuffer.setData1((opcode & 0x0F00) >> 8);
+                fetchDecodeBuffer.setData2((opcode & 0x00F0) >> 4);
                 break;
             case 4:
                 fetchDecodeBuffer.setOpcode(new AddRegister());
-                fetchDecodeBuffer.setData1((short)((opcode & 0x0F00) >> 8));
-                fetchDecodeBuffer.setData2((short)((opcode & 0x00F0) >> 4));
+                fetchDecodeBuffer.setData1((opcode & 0x0F00) >> 8);
+                fetchDecodeBuffer.setData2((opcode & 0x00F0) >> 4);
                 break;
             case 5:
                 fetchDecodeBuffer.setOpcode(new SubRegister());
-                fetchDecodeBuffer.setData1((short)((opcode & 0x0F00) >> 8));
-                fetchDecodeBuffer.setData2((short)((opcode & 0x00F0) >> 4));
+                fetchDecodeBuffer.setData1((opcode & 0x0F00) >> 8);
+                fetchDecodeBuffer.setData2((opcode & 0x00F0) >> 4);
                 break;
             case 6:
                 fetchDecodeBuffer.setOpcode(new ShiftRight());
-                fetchDecodeBuffer.setData1((short)((opcode & 0x0F00) >> 8));
-                fetchDecodeBuffer.setData2((short)((opcode & 0x00F0) >> 4));
+                fetchDecodeBuffer.setData1((opcode & 0x0F00) >> 8);
+                fetchDecodeBuffer.setData2((opcode & 0x00F0) >> 4);
                 break;
             case 7:
                 fetchDecodeBuffer.setOpcode(new SubtractInvertBorrow());
-                fetchDecodeBuffer.setData1((short)((opcode & 0x0F00) >> 8));
-                fetchDecodeBuffer.setData2((short)((opcode & 0x00F0) >> 4));
+                fetchDecodeBuffer.setData1((opcode & 0x0F00) >> 8);
+                fetchDecodeBuffer.setData2((opcode & 0x00F0) >> 4);
                 break;
             case 0xE:
                 fetchDecodeBuffer.setOpcode(new ShiftLeft());
-                fetchDecodeBuffer.setData1((short)((opcode & 0x0F00) >> 8));
-                fetchDecodeBuffer.setData2((short)((opcode & 0x00F0) >> 4));
+                fetchDecodeBuffer.setData1((opcode & 0x0F00) >> 8);
+                fetchDecodeBuffer.setData2((opcode & 0x00F0) >> 4);
                 break;
             default:
                 // XXX Error handling.
@@ -159,80 +159,80 @@ public class FetchDecode {
             break;
         case 9:
             fetchDecodeBuffer.setOpcode(new SkipNextInstructionRegNotEqual());
-            fetchDecodeBuffer.setData1((short)((opcode & 0x0F00) >> 8));
-            fetchDecodeBuffer.setData2((short)((opcode & 0x00F0) >> 4));
+            fetchDecodeBuffer.setData1((opcode & 0x0F00) >> 8);
+            fetchDecodeBuffer.setData2((opcode & 0x00F0) >> 4);
             break;
         case 0xA:
             fetchDecodeBuffer.setOpcode(new SetIToAddress());
-            fetchDecodeBuffer.setData1((short)(opcode & 0x0FFF));
+            fetchDecodeBuffer.setData1(opcode & 0x0FFF);
             break;
         case 0xB:
             fetchDecodeBuffer.setOpcode(new JumpToAddressPlusRegister());
-            fetchDecodeBuffer.setData1((short)(opcode & 0x0FFF));
+            fetchDecodeBuffer.setData1(opcode & 0x0FFF);
             break;
         case 0xC:
             fetchDecodeBuffer.setOpcode(new SetRegisterToRand());
-            fetchDecodeBuffer.setData1((short)((opcode & 0x0F00) >> 8));
-            fetchDecodeBuffer.setData2((short)(opcode & 0x00FF));
+            fetchDecodeBuffer.setData1((opcode & 0x0F00) >> 8);
+            fetchDecodeBuffer.setData2(opcode & 0x00FF);
             break;
         case 0xD:
             fetchDecodeBuffer.setOpcode(new Draw());
-            fetchDecodeBuffer.setData1((short)((opcode & 0x0F00) >> 8));
-            fetchDecodeBuffer.setData2((short)((opcode & 0x00F0) >> 4));
-            fetchDecodeBuffer.setData3((short)(opcode & 0x000F));
+            fetchDecodeBuffer.setData1((opcode & 0x0F00) >> 8);
+            fetchDecodeBuffer.setData2((opcode & 0x00F0) >> 4);
+            fetchDecodeBuffer.setData3(opcode & 0x000F);
             break;
         case 0xE:
             switch (lastNum) {
             case 0x1:
                 fetchDecodeBuffer.setOpcode(new SkipKeyNotEqualRegister());
-                fetchDecodeBuffer.setData1((short)((opcode & 0x0F00) >> 8));
+                fetchDecodeBuffer.setData1((opcode & 0x0F00) >> 8);
                 break;
             case 0xE:
                 fetchDecodeBuffer.setOpcode(new SkipKeyEqualRegister());
-                fetchDecodeBuffer.setData1((short)((opcode & 0x0F00) >> 8));
+                fetchDecodeBuffer.setData1((opcode & 0x0F00) >> 8);
                 break;
             default:
                 // XXX Error handling.
             }
             break;
         case 0xF:
-            short lastBytes = (short)(opcode & 0x00FF);
+            int lastBytes = (opcode & 0x00FF);
             switch (lastBytes) {
             case 0x07:
                 fetchDecodeBuffer.setOpcode(new SetRegisterToDelayTimer());
-                fetchDecodeBuffer.setData1((short)((opcode & 0x0F00) >> 8));
+                fetchDecodeBuffer.setData1((opcode & 0x0F00) >> 8);
                 break;
             case 0x0A:
                 fetchDecodeBuffer.setOpcode(new WaitForKeyPress());
-                fetchDecodeBuffer.setData1((short)((opcode & 0x0F00) >> 8));
+                fetchDecodeBuffer.setData1((opcode & 0x0F00) >> 8);
                 break;
             case 0x15:
                 fetchDecodeBuffer.setOpcode(new SetDelayTimer());
-                fetchDecodeBuffer.setData1((short)((opcode & 0x0F00) >> 8));
+                fetchDecodeBuffer.setData1((opcode & 0x0F00) >> 8);
                 break;
             case 0x18:
                 fetchDecodeBuffer.setOpcode(new SetSoundTimer());
-                fetchDecodeBuffer.setData1((short)((opcode & 0x0F00) >> 8));
+                fetchDecodeBuffer.setData1((opcode & 0x0F00) >> 8);
                 break;
             case 0x1E:
                 fetchDecodeBuffer.setOpcode(new AddRegisterToI());
-                fetchDecodeBuffer.setData1((short)((opcode & 0x0F00) >> 8));
+                fetchDecodeBuffer.setData1((opcode & 0x0F00) >> 8);
                 break;
             case 0x29:
                 fetchDecodeBuffer.setOpcode(new SetIToSpriteLocation());
-                fetchDecodeBuffer.setData1((short)((opcode & 0x0F00) >> 8));
+                fetchDecodeBuffer.setData1((opcode & 0x0F00) >> 8);
                 break;
             case 0x33:
                 fetchDecodeBuffer.setOpcode(new StoreBCDToMemory());
-                fetchDecodeBuffer.setData1((short)((opcode & 0x0F00) >> 8));
+                fetchDecodeBuffer.setData1((opcode & 0x0F00) >> 8);
                 break;
             case 0x55:
                 fetchDecodeBuffer.setOpcode(new VectorStoreRegister());
-                fetchDecodeBuffer.setData1((short)((opcode & 0x0F00) >> 8));
+                fetchDecodeBuffer.setData1((opcode & 0x0F00) >> 8);
                 break;
             case 0x65:
                 fetchDecodeBuffer.setOpcode(new VectorLoadRegister());
-                fetchDecodeBuffer.setData1((short)((opcode & 0x0F00) >> 8));
+                fetchDecodeBuffer.setData1((opcode & 0x0F00) >> 8);
                 break;
             default:
                 // XXX Error handling.
