+++ src/main/java/fortville/opcodes/VectorLoadRegister.java
@@ -10,7 +10,7 @@ import fortville.interfaces.Opcode;
  */
 public class VectorLoadRegister implements Opcode {
     @Override
-    public void execute(short data1, short data2, short data3,
+    public void execute(int data1, int data2, int data3,
         Memory memory, Display display, Registers registers) {
         /*
          * Fx65 - LD Vx, [I]
@@ -18,11 +18,11 @@ public class VectorLoadRegister implements Opcode {
          * The interpreter reads values from memory starting at location I
          * into registers V0 through Vx.
          */
-        short VX = data1;
-        short addrOfI = registers.loadI(); // Get memory address from I.
+        int VX = data1;
+        int addrOfI = registers.loadI();
 
-        for (short regNum = 0; regNum <= VX; regNum++) {
-            byte data = memory.load(addrOfI + regNum);
+        for (int regNum = 0; regNum <= VX; regNum++) {
+            int data = memory.load(addrOfI + regNum);
             registers.storeRegister(regNum, data);
         }
     }
